Address upstream review comments including:
1. cleanup resources in all tests
2. remove MAX_REQ_NUM
3. use common opts framework from msg_pingpong.c
4. remove FI_RMA from mbw_mr
5. don't register memory in mbr_mr
6. set fi_info mode bits correctly
7. address any sync issues
8. convert get_time to return usecs and make common
