+++ ported/osu/ft_utils.c
