+++ ported/osu/ft_utils.h
